# spring-server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Petstore - OpenAPI 3.0
  version: 1.0.11
servers:
- url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /category:
    get:
      description: Returns list of categories
      operationId: getCategoryList
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                type: array
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Category not found
      summary: Get list of all categories
      tags:
      - category
      x-accepts: application/json
      x-tags:
      - tag: category
    post:
      description: Add a new category to the store
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
          application/xml:
            schema:
              $ref: '#/components/schemas/Category'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Category'
        description: Create a new category in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful operation
        "405":
          description: Invalid input
      summary: Add a new category to the store
      tags:
      - category
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: category
    put:
      description: Update an existing category by Id
      operationId: updateCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
          application/xml:
            schema:
              $ref: '#/components/schemas/Category'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Category'
        description: Update an existent category in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Category not found
        "405":
          description: Validation exception
      summary: Update an existing category
      tags:
      - category
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: category
  /category/{categoryId}:
    delete:
      description: delete a category
      operationId: deleteCategory
      parameters:
      - description: Category id to delete
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: successful operation
        "400":
          description: Invalid category value
      summary: Deletes a category
      tags:
      - category
      x-accepts: application/json
      x-tags:
      - tag: category
    get:
      description: Returns a single category
      operationId: getCategory
      parameters:
      - description: ID of category to return
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Category not found
      summary: Find category by ID
      tags:
      - category
      x-accepts: application/json
      x-tags:
      - tag: category
...
components:
...
  schemas:
...
    Category:
      example:
        name: Dog
        id: 1
      properties:
        id:
          example: 1
          format: int64
          type: integer
        name:
          example: Dog
          type: string
      type: object
      xml:
        name: category
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey